feat: removido layout responsivo
