--- lib/responsive/responsive_layout.dart
@@ -1,42 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-
-import '../services/user_provider.dart';
-import '../utils/dimensions.dart';
-
-class ResponsiveLayout extends StatefulWidget {
-  final Widget webScreenLayout;
-  final Widget mobileScreenLayout;
-  const ResponsiveLayout(
-      {Key? key,
-      required this.webScreenLayout,
-      required this.mobileScreenLayout})
-      : super(key: key);
-
-  @override
-  State<ResponsiveLayout> createState() => _ResponsiveLayoutState();
-}
-
-class _ResponsiveLayoutState extends State<ResponsiveLayout> {
-  @override
-  void initState() {
-    super.initState();
-    addData();
-  }
-
-  addData() async {
-    UserProvider _userProvider = Provider.of(context, listen: false);
-    await _userProvider.refreshUser();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(builder: ((context, constraints) {
-      if (constraints.maxWidth > webScreenSize) {
-        return widget.webScreenLayout;
-      } else {
-        return widget.mobileScreenLayout;
-      }
-    }));
-  }
-}
