--- lib/responsive/web_screen_layout.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WebScreenLayout extends StatelessWidget {
-  const WebScreenLayout({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Text("This is Web"),
-      ),
-    );
-  }
-}
