+++ lib/utils/colors.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-const mobileBackgroundColor = Color.fromRGBO(0, 0, 0, 1);
-const mobileHighlightBackgroundColor = Color.fromARGB(255, 25, 25, 25);
+const backgroundColor = Color.fromRGBO(0, 0, 0, 1);
+const highlightBackgroundColor = Color.fromARGB(255, 25, 25, 25);
 const primaryColor = Color.fromRGBO(35, 53, 224, 88);
 const secondaryColor = Color.fromRGBO(67, 130, 250, 98);
 const hightLightColor = Colors.blue;
